move delete-folder to shared components
