+++ app/src/modules/files/routes/collection.vue
@@ -20,7 +20,6 @@ import { useMoveToFolder } from '@/composables/use-move-to-folder';
 import { useCollectionPermissions } from '@/composables/use-permissions';
 import { usePreset } from '@/composables/use-preset';
 import { emitter, Events } from '@/events';
-import DeleteFolderDialog from '@/modules/files/components/delete-folder-dialog.vue';
 import { useFilesStore } from '@/stores/files.js';
 import { useNotificationsStore } from '@/stores/notifications';
 import { useServerStore } from '@/stores/server';
@@ -32,6 +31,7 @@ import { unexpectedError } from '@/utils/unexpected-error';
 import { uploadFiles } from '@/utils/upload-files';
 import { PrivateView, PrivateViewHeaderBarActionButton } from '@/views/private';
 import AddFolder from '@/views/private/components/add-folder.vue';
+import DeleteFolderDialog from '@/views/private/components/delete-folder-dialog.vue';
 import DrawerBatch from '@/views/private/components/drawer-batch.vue';
 import ExportSidebarDetail from '@/views/private/components/export-sidebar-detail.vue';
 import FilesNavigation from '@/views/private/components/files-navigation.vue';
