+++ app/src/views/private/components/delete-folder-dialog.vue
